Trait air_interpreter_sede::Representation

pub trait Representation {
    type SerializeError;
    type DeserializeError;
    type WriteError;
    type Format;
    type SerializedValue: Deref<Target = [u8]>;

    // Required method
    fn get_format(&self) -> Self::Format;
}

A formatter intended for particular type, a base type that defines generic behavior used by particular implementations.

Required Associated Types

type SerializeError

type DeserializeError

type WriteError

type Format

type SerializedValue: Deref<Target = [u8]>

Required Methods

fn get_format(&self) -> Self::Format